Add Cron examples and update README
